Add files via upload
There are approximately 13,000 add-on packages for the programming language R. While discussing with friends and colleagues about a few useful and popular sources of open source I realized it’s not easy to refer to a list that is short in description and handy for a quick reference. I also realized that the terms ‘module’, ‘package’ & ‘library’ which are often loosely used amongst machine learning practitioners are defined differently by different programming languages. To restrict our definition to Python, the most popular ML language, a module “is a .py file that defines one or more function/classes which you intend to reuse in different codes of program”. In Python package refers to “a directory of Python module(s)”. While a library is used “loosely to describe a collection of the core modules”.